+++ example/lib/screens/surface/info_bars.dart
@@ -34,6 +34,35 @@ class _InfoBarsPageState extends State<InfoBarsPage> with PageMixin {
         ),
         subtitle(
           content: const Text(
+            'A InfoBar as a popup',
+          ),
+        ),
+        CardHighlight(
+          backgroundColor: FluentTheme.of(context).micaBackgroundColor,
+          child: Padding(
+            padding: const EdgeInsets.all(8.0),
+            child: Button(
+              onPressed: () {
+                displayInfoBar(context, builder: (context, close) {
+                  return InfoBar(
+                    title: const Text('You can not do that :/'),
+                    content: const Text(
+                        'A proper warning message of why the user can not do that :/'),
+                    action: IconButton(
+                      icon: const Icon(FluentIcons.clear),
+                      onPressed: close,
+                    ),
+                    severity: InfoBarSeverity.warning,
+                  );
+                });
+              },
+              child: const Text('Show InfoBar'),
+            ),
+          ),
+          codeSnippet: '''''',
+        ),
+        subtitle(
+          content: const Text(
             'A closable InfoBar with options to change its severity',
           ),
         ),